add back TemplateBindingSpec
