+++ groovy.header
@@ -13,16 +13,3 @@
     See the License for the specific language governing permissions and
     limitations under the License.
 */
-package org.boozallen.plugins.jte.testcategories
-
-class TestCategories {
-
-}
-
-// should be in other files but ...
-interface Regression{}
-interface Unit{}
-interface Integration{}
-
-interface InProgress{}
-interface Unstable{}
