--- src/test/groovy/org/boozallen/plugins/jte/config/TemplateConfigDslSpec.groovy
@@ -1,365 +1,0 @@
-/*
-    Copyright 2018 Booz Allen Hamilton
-
-    Licensed under the Apache License, Version 2.0 (the "License");
-    you may not use this file except in compliance with the License.
-    You may obtain a copy of the License at
-
-        http://www.apache.org/licenses/LICENSE-2.0
-
-    Unless required by applicable law or agreed to in writing, software
-    distributed under the License is distributed on an "AS IS" BASIS,
-    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-    See the License for the specific language governing permissions and
-    limitations under the License.
-*/
-package org.boozallen.plugins.jte.config
-
-import spock.lang.*
-import org.junit.*
-import org.boozallen.plugins.jte.util.RunUtils
-import org.jenkinsci.plugins.workflow.cps.EnvActionImpl
-import org.jvnet.hudson.test.GroovyJenkinsRule
-
-class TemplateConfigDslSpec extends Specification {
-
-    @Shared
-    @ClassRule
-    @SuppressWarnings('JUnitPublicField')
-    public GroovyJenkinsRule groovyJenkinsRule = new GroovyJenkinsRule()
-
-    def setup(){
-        EnvActionImpl env = Mock()
-        env.getProperty("someField") >> "envProperty"
-
-        GroovySpy(TemplateConfigDsl, global:true)
-        TemplateConfigDsl.getEnvironment() >> env
-
-        GroovySpy(RunUtils, global:true)
-        RunUtils.getClassLoader() >> groovyJenkinsRule.jenkins.getPluginManager().uberClassLoader
-    }
-
-    def "include Jenkins env var in configuration"(){
-        setup:
-        String config = "a = env.someField"
-
-        when:
-        PipelineConfigurationObject configObject = TemplateConfigDsl.parse(config)
-
-        then:
-        configObject.config == [ a: "envProperty" ]
-        configObject.merge.isEmpty()
-        configObject.override.isEmpty()
-    }
-
-    def 'Empty Config File'(){
-        setup:
-        String config = ""
-
-        when:
-        PipelineConfigurationObject configObject = TemplateConfigDsl.parse(config)
-
-        then:
-        configObject.config == [:]
-        configObject.merge.isEmpty()
-        configObject.override.isEmpty()
-    }
-
-    def 'Flat Keys Configuration'(){
-        setup:
-            String config = """
-            a = 3
-            b = "hi"
-            c = true
-            """
-        when:
-            PipelineConfigurationObject configObject = TemplateConfigDsl.parse(config)
-        then:
-            configObject.config == [
-                a: 3,
-                b: "hi",
-                c: true
-            ]
-    }
-
-    def 'Nested Keys Configuration'(){
-        setup:
-            String config = """
-            random = "hi"
-            application_environments{
-                dev{
-                    field = true
-                }
-                test{
-                    field = false
-                }
-            }
-            blah{
-                another{
-                    field = "hey"
-                }
-            }
-            """
-        when:
-            PipelineConfigurationObject configObject = TemplateConfigDsl.parse(config)
-        then:
-            configObject.config == [
-                random: "hi",
-                application_environments: [
-                    dev: [
-                        field: true
-                    ],
-                    test: [
-                        field: false
-                    ]
-                ],
-                blah: [
-                    another: [
-                        field: "hey"
-                    ]
-                ]
-            ]
-    }
-
-    def 'One Merge First Key'(){
-        setup:
-            String config = """
-            application_environments{
-                merge = true
-            }
-            """
-        when:
-            PipelineConfigurationObject configObject = TemplateConfigDsl.parse(config)
-        then:
-            configObject.merge == [ "application_environments" ] as Set
-    }
-
-    def 'One Merge Nested Key'(){
-        setup:
-            String config = """
-            application_environments{
-                dev{
-                    merge = true
-                }
-            }
-            """
-        when:
-            PipelineConfigurationObject configObject = TemplateConfigDsl.parse(config)
-        then:
-            configObject.merge == [ "application_environments.dev" ] as Set
-    }
-
-    def 'Multi-Merge'(){
-        setup:
-            String config = """
-            application_environments{
-                dev{
-                    merge = true
-                }
-                test{
-                    merge = true
-                }
-            }
-            """
-        when:
-            PipelineConfigurationObject configObject = TemplateConfigDsl.parse(config)
-        then:
-            configObject.merge == [ "application_environments.dev", "application_environments.test" ] as Set
-    }
-
-    def 'One Override First Key'(){
-        when:
-            String config = """
-            application_environments{
-                override = true
-            }
-            """
-            PipelineConfigurationObject configObject = TemplateConfigDsl.parse(config)
-        then:
-            configObject.override == [ "application_environments" ] as Set
-    }
-
-    def 'One Override Nested Key'(){
-        when:
-            String config = """
-            application_environments{
-                dev{
-                    override = true
-                }
-            }
-            """
-            PipelineConfigurationObject configObject = TemplateConfigDsl.parse(config)
-        then:
-            configObject.override == [ "application_environments.dev" ] as Set
-    }
-
-    def 'Multi-Override'(){
-        setup:
-            String config = """
-            application_environments{
-                dev{
-                    override = true
-                }
-                test{
-                    override = true
-                }
-            }
-            """
-        when:
-            PipelineConfigurationObject configObject = TemplateConfigDsl.parse(config)
-        then:
-            configObject.override == [ "application_environments.dev", "application_environments.test" ] as Set
-    }
-
-    def 'File Access Throws Security Exception'(){
-        setup:
-            String config = """
-            password = new File("/etc/passwd").text
-            """
-        when:
-            TemplateConfigDsl.parse(config)
-        then:
-            thrown(SecurityException)
-    }
-
-    def "nested blank entry results in empty hashmap"(){
-        setup:
-            String config = """
-            application_environments{
-                dev
-            }
-            """
-        when:
-            PipelineConfigurationObject configObject = TemplateConfigDsl.parse(config)
-        then:
-            configObject.getConfig() == [
-                application_environments: [
-                    dev: [:]
-                ]
-            ]
-    }
-
-    def "root blank entry results in empty hashmap"(){
-        setup:
-            String config = "field"
-        when:
-            PipelineConfigurationObject configObject = TemplateConfigDsl.parse(config)
-        then:
-            configObject.getConfig() == [ field: [:] ]
-    }
-
-    def "merge key when not true is not added to list"(){
-        setup:
-            String config = "a{ merge = false }"
-        when:
-            PipelineConfigurationObject configObject = TemplateConfigDsl.parse(config)
-        then:
-            configObject.getConfig() == [ a: [ merge: false ] ]
-            configObject.merge == [] as Set
-    }
-
-    def "override key when not true is not added to list"(){
-        setup:
-            String config = "a{ override = false }"
-        when:
-            PipelineConfigurationObject configObject = TemplateConfigDsl.parse(config)
-        then:
-            configObject.getConfig() == [ a: [ override: false ] ]
-            configObject.override == [] as Set
-    }
-
-    def "syntax validation for unquoted values"(){
-        setup:
-            String config = "a = b"
-        when:
-            PipelineConfigurationObject configObject = TemplateConfigDsl.parse(config)
-        then:
-            TemplateConfigException ex = thrown()
-            ex.message.contains('did you mean: a = "b"')
-    }
-
-    def "syntax validation for setting configs equal to blocks"(){
-        setup:
-            String config = """
-            a = b{
-                field = true
-            }
-            """
-        when:
-            PipelineConfigurationObject configObject = TemplateConfigDsl.parse(config)
-        then:
-            thrown(TemplateConfigException)
-    }
-
-    def "array lists are appropriately serialized"(){
-        setup:
-            String config = "field = [ 'a', 'b-c' ]"
-            Map expectedConfig = [
-                field: [ "a", "b-c" ]
-            ]
-            def originalConfig, reparsedConfig
-        when:
-            originalConfig = TemplateConfigDsl.parse(config)
-            reparsedConfig = TemplateConfigDsl.parse(TemplateConfigDsl.serialize(originalConfig))
-        then:
-            originalConfig.getConfig() == expectedConfig
-            reparsedConfig.getConfig() == expectedConfig
-    }
-
-    def "Double Quote String block keys with hyphens are appropriately serialized"(){
-        setup:
-            String config = "\"some-block\"{}"
-            Map expectedConfig = [ "some-block": [:] ]
-            def originalConfig, reparsedConfig
-        when:
-            originalConfig = TemplateConfigDsl.parse(config)
-            reparsedConfig = TemplateConfigDsl.parse(TemplateConfigDsl.serialize(originalConfig))
-            println TemplateConfigDsl.serialize(originalConfig)
-        then:
-            originalConfig.getConfig() == expectedConfig
-            reparsedConfig.getConfig() == expectedConfig
-    }
-
-    def "Single Quote String block keys with hyphens are appropriately serialized"(){
-        setup:
-            String config = "'some-block'{}"
-            Map expectedConfig = [ "some-block": [:] ]
-            def originalConfig, reparsedConfig
-        when:
-            originalConfig = TemplateConfigDsl.parse(config)
-            reparsedConfig = TemplateConfigDsl.parse(TemplateConfigDsl.serialize(originalConfig))
-            println TemplateConfigDsl.serialize(originalConfig)
-        then:
-            originalConfig.getConfig() == expectedConfig
-            reparsedConfig.getConfig() == expectedConfig
-    }
-
-    def "Single Quote String block keys with / are appropriately serialized"(){
-        setup:
-            String config = "'some/block'{}"
-            Map expectedConfig = [ "some/block": [:] ]
-            def originalConfig, reparsedConfig
-        when:
-            originalConfig = TemplateConfigDsl.parse(config)
-            reparsedConfig = TemplateConfigDsl.parse(TemplateConfigDsl.serialize(originalConfig))
-            println TemplateConfigDsl.serialize(originalConfig)
-        then:
-            originalConfig.getConfig() == expectedConfig
-            reparsedConfig.getConfig() == expectedConfig
-    }
-
-    def "Double Quote String block keys with / are appropriately serialized"(){
-        setup:
-            String config = "\"some/block\"{}"
-            Map expectedConfig = [ "some/block": [:] ]
-            def originalConfig, reparsedConfig
-        when:
-            originalConfig = TemplateConfigDsl.parse(config)
-            reparsedConfig = TemplateConfigDsl.parse(TemplateConfigDsl.serialize(originalConfig))
-            println TemplateConfigDsl.serialize(originalConfig)
-        then:
-            originalConfig.getConfig() == expectedConfig
-            reparsedConfig.getConfig() == expectedConfig
-    }
-
-}
